[lp-2189] : create filings package and move classes into it
